Split up ArchiveTree
