+++ cardigan/stories/components/ArchiveTree/ArchiveTree.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Canvas, Story, ArgsTable } from '@storybook/addon-docs/blocks';
 import * as stories from './ArchiveTree.stories.tsx';
-import ArchiveTree from '@weco/catalogue/components/ArchiveTree/ArchiveTree';
+import ArchiveTree from '@weco/catalogue/components/ArchiveTree';
 
 <Meta title={`Components/ArchiveTree`} component={ArchiveTree} />
