+++ cardigan/stories/components/ArchiveTree/ArchiveTree.stories.tsx
@@ -1,4 +1,4 @@
-import ArchiveTree from '@weco/catalogue/components/ArchiveTree/ArchiveTree';
+import ArchiveTree from '@weco/catalogue/components/ArchiveTree';
 import collectionTree from '@weco/cardigan/stories/data/collection-tree';
 
 const Template = args => <ArchiveTree {...args} />;
